Changes to be able to deploy run in Glassfish. Changed a few properties from Integer to Long in Auction because of Derby. Realm needed Jaxb annotations because of Moxy, JSON marshaling and also dedicated getters and setters. EclipseLink does not support Java 8 streams, workaround convert IndirectList to ArrayList.

+++ batch/src/main/java/com/radcortez/wow/auctions/batch/process/data/AuctionDataItemReader.java
@@ -93,10 +93,10 @@ public class AuctionDataItemReader extends AbstractAuctionFileProcess implements
             auction.setOwnerRealm(parser.getString());
             parser.next();
             parser.next();
-            auction.setBid(parser.getInt());
+            auction.setBid(parser.getLong());
             parser.next();
             parser.next();
-            auction.setBuyout(parser.getInt());
+            auction.setBuyout(parser.getLong());
             parser.next();
             parser.next();
             auction.setQuantity(parser.getInt());
